Hardcode provider to /etc/puppetlabs/puppet/ssl
This commit updates the SSL configuration for the connection used manage
directory settings to use certificates and keys from /etc/puppetlabs/puppet/ssl.
Previously, the connection would try to use `Puppet.settings` to discover
this configuration dynamically. The problem is that Bolt `apply` uses
a disposable directory under `/tmp` that does not contain copies of the
SSL keys and cert and thus fails to apply `puppet_ds` resources.

There is likely a better solution, but that will require hacking on the
ResourceApi a bit to thread a parameter in. /etc/puppetlabs/puppet/ssl is
the right directory 99.9% of the time, so hardcoding will work in most cases.